+++ src/utils/validators.rs
@@ -1,4 +1,4 @@
-pub fn is_positive<T: PartialOrd<f64>>(value: T) -> bool {
+pub fn is_positive<T: PartialOrd<f32>>(value: T) -> bool {
     if value > 0.0 {
         return true;
     }
